IKASAN-293:File transfer connector bug. Minimize failures on doCommit

+++ ikasaneip-0.7.x/connector-basefiletransfer/src/main/java/org/ikasan/connector/basefiletransfer/outbound/command/DeliverFileCommand.java
@@ -249,19 +249,27 @@ public class DeliverFileCommand extends AbstractBaseFileTransferTransactionalRes
             changeDirectory = true;
         }
 
-        if (this.fileExists(this.fileName))
+        try
         {
-            if (this.overwriteExisting)
+            /*
+             * The extra delete operation is added to cater for the SFTP rename behavior:
+             * the existing file must be first removed, before delivering another with the
+             * same name.
+             * 
+             * The case were we are trying to deliver a file that already exists on target system
+             * and overwrite flag is false is handled at the execute stage; an exception would be thrown
+             * and we never reach this stage.
+             */
+            if (this.overwriteExisting && this.fileExists(this.fileName))
             {
                 logger.debug("Deleting existing file of the same name as the one we are delivering [" //$NON-NLS-1$
-                    + this.fileName + "]");
+                        + this.fileName + "]");
                 deleteFile(this.fileName);
             }
-            else
-            {
-                throw new ResourceException("Cannot rename file as it a file of the destination name already exists ["
-                        + this.tempFileName + "]");
-            }
+        }
+        catch (ResourceException e)
+        {
+            logger.warn(e);
         }
 
         renameFile(tempFileName, fileName);
